Crypto_NewsApi version2
The app links to the webpage shown in the list.
The QueryUtils was cleaned up, with mURL added to CryptoInfo
Url response link updated to find crypto information.

Things to be resolved. Need to clean up the setting preference.
Also need to clean up the orderBy function, it only finds stories for the current date.