Crypto_NewsApi v1 - June 2020
The app links to the webpage shown in the list.

"If you fork this project you need to add an API key from the NewApi site. "
Things to be resolved. Need to clean up the setting preference.
Also need to clean up the orderBy function, it only finds stories for the current date.